fix tests and create separate constructor file
