+++ lib/2dobox.js
@@ -1,6 +1,9 @@
 /*jshint esversion: 6 */
+
+//Have submit button disabled = true at start;
 const $ = require('jquery');
 
+const CreateIdea = require('./constructor');
 const title = $('.title-field');
 const body = $('.body-field');
 const search = $('.search-field');
@@ -189,13 +192,13 @@ body.on('keyup', function(){
   }
 });
 
-function CreateIdea(title, body, id, quality) {
-  this.title = title;
-  this.body = body;
-  this.id = id || Date.now();
-  this.quality = quality || "Normal";
-  this.done = false;
-}
+// function CreateIdea(title, body, id, quality) {
+//   this.title = title;
+//   this.body = body;
+//   this.id = id || Date.now();
+//   this.quality = quality || "Normal";
+//   this.done = false;
+// }
 
 function displayIdea(titleInput, bodyInput, id, quality){
   $('.idea-list').prepend(
