+++ lib/index.js
@@ -2,6 +2,7 @@
 
 import './2dobox';
 import './quality';
+import './constructor';
